+++ fractal_server/app/routes/api/v2/job.py
@@ -8,6 +8,7 @@ from fastapi import status
 from fastapi.responses import StreamingResponse
 from sqlmodel import select
 
+from .....zip_tools import _zip_folder_to_byte_stream_iterator
 from ....db import AsyncSession
 from ....db import get_async_db
 from ....models.v2 import JobV2
@@ -18,7 +19,6 @@ from ....schemas.v2 import JobStatusTypeV2
 from ....security import current_active_user
 from ....security import User
 from ...aux._job import _write_shutdown_file
-from ...aux._job import _zip_folder_to_byte_stream
 from ...aux._runner import _check_shutdown_is_supported
 from ._aux_functions import _get_job_check_owner
 from ._aux_functions import _get_project_check_owner
@@ -118,7 +118,7 @@ async def download_job_logs(
     db: AsyncSession = Depends(get_async_db),
 ) -> StreamingResponse:
     """
-    Download job folder
+    Download zipped job folder
     """
     output = await _get_job_check_owner(
         project_id=project_id,
@@ -127,15 +127,11 @@ async def download_job_logs(
         db=db,
     )
     job = output["job"]
-
-    # Create and return byte stream for zipped log folder
-    PREFIX_ZIP = Path(job.working_dir).name
-    zip_filename = f"{PREFIX_ZIP}_archive.zip"
-    byte_stream = _zip_folder_to_byte_stream(folder=job.working_dir)
+    zip_name = f"{Path(job.working_dir).name}_archive.zip"
     return StreamingResponse(
-        iter([byte_stream.getvalue()]),
+        _zip_folder_to_byte_stream_iterator(folder=job.working_dir),
         media_type="application/x-zip-compressed",
-        headers={"Content-Disposition": f"attachment;filename={zip_filename}"},
+        headers={"Content-Disposition": f"attachment;filename={zip_name}"},
     )